# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Payments Processing API
  description: >
    REST API for managing the full lifecycle of financial payments —
    creation, validation, transmission, completion/failure — with a
    complete audit trail of status transitions. The single operator owns
    multiple accounts (INR/USD); a KPI/analytics dashboard and API rate
    limiting are provided per the 31-Jul-2026 customer meeting.
  version: "1.1.0"
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
  /accounts:
    get:
      tags: [Accounts]
      summary: List the operator's accounts
      operationId: listAccounts
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /accounts/{id}:
    get:
      tags: [Accounts]
      summary: Get a single account by ID
      operationId: getAccountById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
...
  responses:
    RateLimited:
      description: Rate limit exceeded (token bucket exhausted)
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the client may retry
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    CurrencyCode:
      type: string
      enum: [INR, USD]
      description: >
        Only INR and USD are supported (customer-confirmed 31 Jul 2026).
        No cross-currency conversion is performed.
    AccountStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b2c3d4e5-1111-4a11-8a11-111111111111"
        label:
          type: string
          example: "Primary INR Savings"
        accountNumber:
          type: string
          example: ACC1000001
        currency:
          $ref: '#/components/schemas/CurrencyCode'
        status:
          $ref: '#/components/schemas/AccountStatus'
...
    ApiError:
      type: object
      properties:
        errorCode:
          type: string
          example: INVALID_AMOUNT
        message:
          type: string
          example: Amount must be greater than 0
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /api/v1/payments